mcmc.sh seems to work
a set of experiments is now classified as DATA_IS_ABSOLUTE if after loading from hdf5 all experiments provide absolute data; the earlier import method for data, before hdf5, forced the normalisation method to be the same for all experiments; this was changed at the same time as hdf5 import was introduced.

Added a Makefile for PDC/Beskow

updated ubuntu Makefile

the sbtab import (SBtab to HDF5) feature now respects the !Scale of given parameter columns.